chore(money): thread CentsAmount brand across the api/web boundary
Move the CentsAmount brand, cents(), and unsafeCents() into @stll/money
so a single nominal brand is shared by backend and frontend. Brand the
prorateHourlyCents/applyMarkupCents helper inputs and returns so minor-unit
amounts stay typed end to end (Drizzle column -> Eden -> browser preview).

apps/api/src/lib/money.ts re-exports from @stll/money to keep API-side
consumers and the Drizzle schema unchanged.
